requirements.txt
allure-pytest==2.8.13
json5==0.9.4
lxml==4.5.0
PyMySQL==0.9.3
pytest==6.0.2
PyYAML==5.3.1
requests==2.23.0
selenium==3.141.0
split==0.4
urllib3==1.25.8
xlrd==1.2.0